refactor(app-ui): 优化 UI 代码结构
