+++ app/src/main/java/yangfentuozi/runner/app/ui/activity/ExecShortcutActivity.kt
@@ -5,7 +5,7 @@ import android.os.PersistableBundle
 import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import yangfentuozi.runner.app.ui.dialogs.ExecDialog
+import yangfentuozi.runner.app.ui.components.ExecDialog
 import yangfentuozi.runner.app.ui.theme.RunnerTheme
 import yangfentuozi.runner.shared.data.CommandInfo
