+++ app/src/main/java/yangfentuozi/runner/app/ui/activity/MainActivity.kt
@@ -3,7 +3,7 @@ package yangfentuozi.runner.app.ui.activity
 import android.os.Bundle
 import androidx.activity.compose.setContent
 import yangfentuozi.runner.app.base.BaseActivity
-import yangfentuozi.runner.app.ui.MainScreen
+import yangfentuozi.runner.app.ui.screens.main.MainScreen
 import yangfentuozi.runner.app.ui.theme.RunnerTheme
 
 class MainActivity : BaseActivity() {
